Use default manifest filename
Both webpack-assets-manifest and external_asset_pipeline assume a
default manifest filename of `manifest.json`. The only reason we weren't
using this default before was because we had been using the StatsPlugin
to produce a different `manifest.json` file. Now that we've removed the
StatsPlugin, we can rely on the default webpack-assets-manifest and
external_asset_pipeline configuration.